requirements/base.txt
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-decouple==3.8
ShopifyAPI==12.3.0
